% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth.R
\name{ytr_oauth}
\alias{ytr_oauth}
\title{Authenticate}
\usage{
ytr_oauth(client.id, client.secret, scope = "yt-analytics.readonly")
}
\arguments{
\item{client.id}{Your \code{client.id}. See details below.}

\item{client.secret}{Your \code{client.secret}. See details below.}

\item{scope}{Scope of token, defaults to \code{NULL}. See details below.}
}
\description{
Authenticate
}
\details{
To get your client.id and client.secret please follow the instructions
\href{https://developers.google.com/youtube/registering_an_application}{here}
Also set the redirect Authorized redirect URIs to your \code{httr::oauth_callback()} URL,
generally your localhost.

Valid scopes:
\itemize{
  \item{\code{yt-analytics-monetary.readonly}}
  \item{\code{yt-analytics.readonly}}
}

See \href{https://developers.google.com/youtube/reporting/guides/authorization/client-side-web-apps}{official documentation}
for more information on the scopes.
}
