% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{ytr_parse}
\alias{ytr_parse}
\title{Parse results}
\usage{
ytr_parse(ytr)
}
\arguments{
\item{ytr}{An object as returned by an API call.}
}
\value{
\code{tibble}
}
\description{
Build data frame from results
}
