% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_jobs.R
\name{ytr_reports}
\alias{ytr_reports}
\title{List jobs}
\usage{
ytr_reports(token, start = Sys.Date() - 3, end = Sys.Date(),
  ids = "channel==MINE", metrics = c("views", "likes"),
  dimensions = NULL, ...)
}
\arguments{
\item{token}{your token as returned by \code{\link{ytr_oauth}}.}

\item{start, end}{Start and end date of class \code{Date}.}

\item{ids}{Ids of chennels or users data to return.}

\item{metrics}{Metrics to return, see\href{https://developers.google.com/youtube/analytics/metrics}{docs}.}

\item{dimensions}{A \code{list} or \code{vector} of YouTube Analytics dimensions, such as \code{video} or \code{ageGroup}, see\href{https://developers.google.com/youtube/reporting/}{docs}.}

\item{...}{Any other parameter to pass to query.}
}
\description{
List jobs
}
\examples{
\dontrun{
token <- ytr_oauth(
  "something.apps.googleusercontent.com",
  "xxXx0X0x0X0xXX0")

ytr <- ytr_reports(token) \%>\%
  ytr_parse()
}

}
\seealso{
\href{https://developers.google.com/youtube/analytics/}{official documentation}
}
